Added handler for CommandManagerStateEvents which catch Train events.
While first Train<unit> or Morph<unit> CommandEvent was showing up
in the events list, subsequent duplicate commands were not being
caught.  That is because, to save space in the binary file, they
were coded as COmmandManagerStateEvents.

This branch has been rebased on updated upstream which exposes these
new events, and a handler for the events has been added.
CommandEvents are cached as they come in and the new handlers
repeats the cached event after replacing the frame timestamp with
it's own.

Still to do:
  - Simulate building queues: wait time, max capacity
  - Apply this change to Zerg (queens, morph events)